moves type functions into separate folder; writes test for getDishTypes
