+++ libs/menu-matriarch/types/src/lib/dish.interface.ts
@@ -28,7 +28,3 @@ export interface FilteredDishesGroup {
   dishes: Dish[];
   placeholderText?: string;
 }
-
-export function getDishTypes(): readonly DishType[] {
-  return ['main', 'side', 'dessert'];
-}
