+++ libs/menu-matriarch/types/src/lib/functions/create-dtos.ts
@@ -1,8 +1,8 @@
-import { DishDto } from "./dish.interface";
-import { MealDto } from "./meal.interface";
-import { MenuDto } from "./menu.interface";
-import { TagDto } from "./tag.interface";
-import { UserDto } from "./user.interface";
+import { DishDto } from "../dish.interface";
+import { MealDto } from "../meal.interface";
+import { MenuDto } from "../menu.interface";
+import { TagDto } from "../tag.interface";
+import { UserDto } from "../user.interface";
 
 export function createUserDto({
   uid,
