+++ libs/menu-matriarch/types/src/lib/functions/get-days.ts
@@ -1,4 +1,4 @@
-import { Day } from "./day.type";
+import { Day } from "../day.type";
 
 export function getDays(startDay: Day = 'Monday'): readonly Day[] {
   switch (startDay) {
