+++ libs/menu-matriarch/types/src/lib/functions/get-dish-types.spec.ts
@@ -1,0 +1,7 @@
+import { getDishTypes } from './get-dish-types';
+
+describe('getDishTypes', () => {
+  it('returns dish types in correct order', () => {
+    expect(getDishTypes()).toEqual(['main', 'side', 'dessert']);
+  });
+});
