+++ libs/menu-matriarch/types/src/lib/functions/get-dish-types.ts
@@ -1,0 +1,5 @@
+import { DishType } from "../dish.interface";
+
+export function getDishTypes(): readonly DishType[] {
+  return ['main', 'side', 'dessert'];
+}
